Fix tz conversion
The `mode` return value can differ depending of pandas versions. Prior `1.5.1` it was returning
a tz naive timestamp therefore requiring the use of `tz_localize` and after it returned a tz aware timestamp
requiring the use of `tz_convert`. This was a bug since `.dt.tz is not None` check was checking the tz of the column but not the mode timestamp.